requirements.txt
# ── Web Framework ─────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
aiofiles==24.1.0

# ── AI / LangChain ────────────────────────────────────────────
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-core==0.3.29
google-generativeai==0.8.3

# ── PDF Processing ────────────────────────────────────────────
pdfplumber==0.11.4
pypdf==5.1.0
camelot-py[cv]==0.11.0
reportlab==4.2.5
pdf2image==1.17.0
Pillow==11.0.0

# ── OCR Engines ───────────────────────────────────────────────
easyocr==1.7.2
paddleocr==2.9.1
paddlepaddle==2.6.2
pytesseract==0.3.13

# ── Image Processing (OpenCV) ─────────────────────────────────
opencv-python-headless==4.10.0.84
numpy==1.26.4
scikit-image==0.24.0
scipy==1.14.1

# ── Data Processing ───────────────────────────────────────────
pandas==2.2.3
openpyxl==3.1.5
pydantic==2.10.3

# ── Database ──────────────────────────────────────────────────
sqlalchemy==2.0.36
aiosqlite==0.20.0

# ── Utilities ─────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.28.1
python-magic==0.4.27
filetype==1.2.0
loguru==0.7.3
